Split intarget tools into a separate module
